backend/requirements.txt
Flask>=3.0,<4.0
Flask-Cors>=4.0,<5.0
Flask-Migrate>=4.0,<5.0
Flask-SQLAlchemy>=3.1,<4.0
python-dotenv>=1.0,<2.0
PyMySQL>=1.1,<2.0
pydantic>=2.7,<3.0
email-validator>=2.0,<3.0
opencv-python-headless>=4.10,<5.0
mediapipe>=0.10,<0.11
